requirements.txt
beaker >= 1.8
bottle >= 0.12
bottle-cork >= 0.12
Bottle-Sqlalchemy >= 0.4
pycrypto >= 2.6
rsa >= 3.4
sqlalchemy >= 1.0

